Add a placeholder titlebar / taskbar icon
Sometimes I use the taskbar to click on icons and with an icon-less
application running it makes it ~50/50 whether the right app is focused.

The placeholder icon is programmer art of a D with a "next track" icon
inside - just to make it more clear than a "play" icon that just looks
like the inside of the D. I thought about doing something with the Rad
logo, but in the end decided to do something simple and let someone with
more talent update the icon.

As for updating the icon, simply run (requires png2ico)

```
data\icon\build.bat
```

and then rebuild `raddbg.exe`

```
build.bat raddbg
```

Tested with VS and clang builds.