[Metagen Linux] Added: More missing Linux implimentations, metagen compiles and runs
I haven't verified how well much of the code works but the one thing metagen
uses, the OS filter iterator, seems happy and I checked a good few of the values
it spat out. The only thing left is the condition variable broadcast and a few
long winded things I don't know how to compelte the implementation on. I don't
really know what a condition variable is or what a broadcast for it is so moving
to testing other things first.

Also the OS_SystemPath_ModuleLoad is not done.

Maintenance: Sorted include lines- because nice.

[Build Batch] Fixed: Exiting with fail status (1) on successful compile instead of success