removed dead codepaths related to ALT key during WM_CHAR events
It appears that keyboard input with the ALT key held does not trigger
WM_CHAR events (only WM_SYSCHAR events).

Moreover, the win32 documentation states that bit 29 (zero-based
indexing) is used to indicate if ALT is held. This would correspond to `bit30`, not `bit29`, as these constant names use one-based indexing.
This makes it clear that the conditions in the deleted code were never
hit, in any circumstance.